+++ stag_lib/Graph/spectrum.cpp
