+++ stag_lib/KDE/data.cpp
