+++ test/cluster_test.cpp
@@ -8,11 +8,11 @@
 #include <stdexcept>
 #include <algorithm>
 #include <gtest/gtest.h>
-#include <cluster.h>
-#include <graph.h>
-#include <random.h>
+#include "Cluster/cluster.h"
+#include "Graph/graph.h"
+#include "Graph/random.h"
 #include <utility.h>
-#include <graphio.h>
+#include "Graph/graphio.h"
 
 // Define some helper test assertions.
 #define EXPECT_FLOATS_NEARLY_EQ(expected, actual, thresh) \
