+++ test/data_test.cpp
@@ -8,10 +8,10 @@
 #include <iostream>
 #include <gtest/gtest.h>
 #include <filesystem>
-#include "data.h"
+#include "KDE/data.h"
 #include "utility.h"
-#include "graph.h"
-#include "random.h"
+#include "Graph/graph.h"
+#include "Graph/random.h"
 
 // Note: for many of these tests, we re-use the edgelist and adjacencylist files
 // however, in general the load_matrix and save_matrix methods should be used
