+++ test/lsh_test.cpp
@@ -8,7 +8,7 @@
 #include <iostream>
 #include <stdexcept>
 #include <gtest/gtest.h>
-#include <lsh.h>
+#include "LSH/lsh.h"
 
 //------------------------------------------------------------------------------
 // Tests for the DataPoint class.
